Fix crash when using hotkey on inventory background (fixes #1)
Note that "background" here refers to the gray area of the inventory GUI,
including the area with the player character, and not the translucent
black layer covering the entire screen.

This is accomplished by checking for any slot ID less than 0, which will
never be a valid inventory slot and will always crash. Prior to this
commit, the slot ID -999 (referring to the black layer) was checked for
specifically.